Move MethodExtensions to geotrellis-util
