+++ raster/src/main/scala/geotrellis/raster/crop/CropMethods.scala
@@ -2,6 +2,8 @@ package geotrellis.raster.crop
 
 import geotrellis.vector._
 import geotrellis.raster._
+import geotrellis.util.MethodExtensions
+
 
 trait CropMethods[T] extends MethodExtensions[T] {
   import Crop.Options
