+++ raster/src/main/scala/geotrellis/raster/mapalgebra/local/Or.scala
@@ -17,6 +17,8 @@
 package geotrellis.raster.mapalgebra.local
 
 import geotrellis.raster._
+import geotrellis.util.MethodExtensions
+
 
 /**
  * Or's cell values of rasters or Int values.
