+++ raster/src/main/scala/geotrellis/raster/mapalgebra/local/Pow.scala
@@ -17,6 +17,8 @@
 package geotrellis.raster.mapalgebra.local
 
 import geotrellis.raster._
+import geotrellis.util.MethodExtensions
+
 
 /**
  * Pows values.
