+++ raster/src/main/scala/geotrellis/raster/mapalgebra/local/Xor.scala
@@ -17,6 +17,8 @@
 package geotrellis.raster.mapalgebra.local
 
 import geotrellis.raster._
+import geotrellis.util.MethodExtensions
+
 
 /**
  * Xor's cell values of rasters or Int values.
