+++ raster/src/main/scala/geotrellis/raster/mapalgebra/zonal/ZonalMethods.scala
@@ -3,6 +3,8 @@ package geotrellis.raster.mapalgebra.zonal
 import geotrellis.raster._
 import geotrellis.raster.summary._
 import geotrellis.raster.histogram._
+import geotrellis.util.MethodExtensions
+
 
 trait ZonalMethods extends MethodExtensions[Tile] {
   def zonalHistogramInt(zones: Tile): Map[Int, Histogram[Int]] =
