+++ raster/src/main/scala/geotrellis/raster/merge/TileMergeMethods.scala
@@ -3,6 +3,8 @@ package geotrellis.raster.merge
 import geotrellis.raster._
 import geotrellis.raster.resample.{NearestNeighbor, ResampleMethod}
 import geotrellis.vector.Extent
+import geotrellis.util.MethodExtensions
+
 
 trait TileMergeMethods[T] extends MethodExtensions[T] {
   def merge(other: T): T
