+++ raster/src/main/scala/geotrellis/raster/rasterize/FeatureRasterizeMethods.scala
@@ -19,6 +19,7 @@ package geotrellis.raster.rasterize
 import geotrellis.raster._
 import geotrellis.raster.rasterize.Rasterize.Options
 import geotrellis.vector.{Geometry,Feature}
+import geotrellis.util.MethodExtensions
 
 
 trait FeatureIntRasterizeMethods[+G <: Geometry, T <: Feature[G,Int]] extends MethodExtensions[T] {
