+++ raster/src/main/scala/geotrellis/raster/render/ColorMethods.scala
@@ -4,10 +4,12 @@ import geotrellis.raster._
 import geotrellis.raster.render.png._
 import geotrellis.raster.histogram.Histogram
 import geotrellis.raster.summary._
+import geotrellis.util.MethodExtensions
 
 import java.awt.image.BufferedImage
 
 import spire.syntax.cfor._
+
 
 trait ColorMethods extends MethodExtensions[Tile] {
   def color(breaksToColors: Map[Int, Int]): Tile =
