+++ raster/src/main/scala/geotrellis/raster/resample/ResampleMethods.scala
@@ -2,8 +2,10 @@ package geotrellis.raster.resample
 
 import geotrellis.raster._
 import geotrellis.vector.Extent
+import geotrellis.util.MethodExtensions
 
 import spire.syntax.cfor._
+
 
 trait ResampleMethods[T <: CellGrid] extends MethodExtensions[T] {
   def resample(extent: Extent, targetExtent: RasterExtent, method: ResampleMethod): T
