+++ raster/src/main/scala/geotrellis/raster/summary/SummaryMethods.scala
@@ -2,6 +2,8 @@ package geotrellis.raster.summary
 
 import geotrellis.raster._
 import geotrellis.raster.histogram._
+import geotrellis.util.MethodExtensions
+
 
 trait SummaryMethods extends MethodExtensions[Tile] {
   /**
