+++ util/src/main/scala/geotrellis/util/MethodExtensions.scala
@@ -1,4 +1,4 @@
-package geotrellis.raster
+package geotrellis.util
 
 trait MethodExtensions[+T] extends Serializable {
   def self: T
